rand in range, correcting unit vectors in grad grid

+++ graphs/maps/generator.h
@@ -5,6 +5,6 @@ int **generate_map(const size_t res_x,
 		   const int min_cost,
 		   const int max_cost);
 
-int ***init_grad_grid(const size_t res_x,
+double ***init_grad_grid(const size_t res_x,
 		      const size_t res_y);
 #endif /* ifndef GRAPHS_MAPS_GENERATOR_H_ */
